cli.py: avoid disabling ssl verify
Currently cli args allow options to ignore ssl-proxy environment
variables and also specify to underlying urllib to skip ssl verification
but also providing a means for users to define these options via
environment variables NDL_SKIP_PROXY and so on. However, this behaviour
is always false and results in disabling ssl verification when we didn't
intend to disable ssl.

Set args to None to indicate no option was specified.  When the switch
is defined, we store 'true' or 'false' appropriately and set the user
preference.  Remove env var parsing as it's not necessary to override or
useful to alternatively define.

Signed-off-by: Jamie Couture <jamie.couture@nasdaq.com>