cli.py: add failed url retry logic
Currently, if Retry is exhausted or requests runs into an exception we
do not track the list of failed files.  Callers are forced to start all
over again and download everything.

When an exception is thrown, add it to the global list of failed urls.

After the first pass of downloads is complete, keep checking the failed
urls list and re-try until it's empty.

Signed-off-by: Jamie Couture <jamie.couture@nasdaq.com>